+++ src/panels/Workflows/LearningMap/__Separator/LearningMap__Separator.scss
